requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.5
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
discord.py==1.6.0
idna==2.10
multidict==5.1.0
numpy==1.20.1
pandas==1.2.2
praw==7.2.0
prawcore==2.0.0
pycodestyle==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
toml==0.10.2
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.3
websocket-client==0.57.0
yarl==1.6.3
